backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlmodel==0.0.14
pymysql==1.1.0
cryptography==41.0.7
alembic==1.12.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-ldap3==2.9.1
python-multipart==0.0.6

# Caching and Session
redis==5.0.1
aioredis==2.0.1

# HTTP Client
httpx==0.25.2
requests==2.31.0

# File Handling
python-magic==0.4.27
Pillow==10.1.0
aiofiles==23.2.1

# Configuration and Environment
python-decouple==3.8
pydantic-settings==2.0.3

# Logging and Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Date and Time
pytz==2023.3

# Email
emails==0.6

# Excel and CSV Processing
pandas==2.1.3
openpyxl==3.1.2

# Task Queue (for background jobs)
celery[redis]==5.3.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1